menulo/priv/menulo/requirements.txt
asgiref==3.2.10
boto3==1.14.59
botocore==1.17.59
dj-database-url==0.5.0
Django==3.1.1
django-autoslug==1.9.8
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
django-storages==1.10
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
numpy==1.19.1
Pillow==7.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.10
whitenoise==5.2.0
